Move dynamic registry tests into their own class for readibility
