+++ fabric-registry-sync-v0/src/testmod/resources/fabric.mod.json
@@ -10,10 +10,11 @@
   },
   "entrypoints": {
     "main": [
+      "net.fabricmc.fabric.test.registry.sync.CustomDynamicRegistryTest",
       "net.fabricmc.fabric.test.registry.sync.RegistrySyncTest"
     ],
     "client": [
-      "net.fabricmc.fabric.test.registry.sync.client.RegistrySyncClientTest"
+      "net.fabricmc.fabric.test.registry.sync.client.DynamicRegistryClientTest"
     ]
   }
 }
